requirements.txt
asgiref==3.7.2
Django==4.2.4
Pillow==10.0.0
sqlparse==0.4.4
stepic==0.5.0
typing_extensions==4.7.1
tzdata==2023.3
